find mesaage by Id in the Repository

+++ src/main/java/com/example/mygreetingapp/repository/IGreetingRepo.java
@@ -3,5 +3,5 @@ package com.example.mygreetingapp.repository;
 import com.example.mygreetingapp.model.Greeting;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IGreetingRepo extends JpaRepository<Greeting,Long> {
+public interface IGreetingRepo extends JpaRepository<Greeting,Integer> {
 }
